alu ops addition

+++ GBA-emulator/cplusplusRewrite/Shifts.cpp
@@ -5,7 +5,7 @@
 
 using namespace shifts;
 
-void Lsl::calcConditions(CPSR_t& cpsr, const uint32_t result, const uint32_t sourceValue, const uint8_t shiftAmount) {
+void Lsl::calcConditions(CPSR_t& cpsr, const uint32_t result, const uint32_t sourceValue, const uint32_t shiftAmount) {
 	if (shiftAmount > 32)
 		cpsr.carry = 0;
 	else if (shiftAmount > 0)
@@ -14,13 +14,13 @@ void Lsl::calcConditions(CPSR_t& cpsr, const uint32_t result, const uint32_t sou
 	cpsr.zero = result == 0;
 }
 
-uint32_t Lsl::shift(uint32_t sourceValue, uint8_t shiftAmount) {
+uint32_t Lsl::shift(uint32_t sourceValue, uint32_t shiftAmount) {
 	uint64_t tmp = sourceValue;
 	return static_cast<uint32_t>(tmp << shiftAmount);
 }
 
 
-void Lsr::calcConditions(CPSR_t& cpsr, const uint32_t result, const uint32_t sourceValue, const uint8_t shiftAmount) {
+void Lsr::calcConditions(CPSR_t& cpsr, const uint32_t result, const uint32_t sourceValue, const uint32_t shiftAmount) {
 	uint64_t tmp = sourceValue;
 	if (shiftAmount > 0)
 		cpsr.carry = (tmp >> (shiftAmount - 1) & 1);
@@ -28,12 +28,12 @@ void Lsr::calcConditions(CPSR_t& cpsr, const uint32_t result, const uint32_t sou
 	cpsr.zero = result == 0;
 }
 
-uint32_t Lsr::shift(const uint32_t sourceValue, const uint8_t shiftAmount) {
+uint32_t Lsr::shift(const uint32_t sourceValue, const uint32_t shiftAmount) {
 	return static_cast<uint64_t>(sourceValue) >> shiftAmount;
 }
 
 
-void Asr::calcConditions(CPSR_t& cpsr, const uint32_t result, const uint32_t sourceValue, const uint8_t shiftAmount) {
+void Asr::calcConditions(CPSR_t& cpsr, const uint32_t result, const uint32_t sourceValue, const uint32_t shiftAmount) {
 	int64_t tmp = static_cast<int64_t>(static_cast<int32_t>(sourceValue));
 	if (shiftAmount != 0)
 		cpsr.carry = (tmp >> (shiftAmount - 1) & 1);
@@ -41,18 +41,18 @@ void Asr::calcConditions(CPSR_t& cpsr, const uint32_t result, const uint32_t sou
 	cpsr.negative = static_cast<int32_t>(result) < 0;
 }
 
-uint32_t Asr::shift(const uint32_t sourceValue, const uint8_t shiftAmount) {
+uint32_t Asr::shift(const uint32_t sourceValue, const uint32_t shiftAmount) {
 	return static_cast<int64_t>(static_cast<int32_t>(sourceValue)) >> shiftAmount;
 }
 
-void Ror::calcConditions(CPSR_t& cpsr, const uint32_t result, const uint32_t sourceValue, const uint8_t shiftAmount) {
+void Ror::calcConditions(CPSR_t& cpsr, const uint32_t result, const uint32_t sourceValue, const uint32_t shiftAmount) {
 	if (shiftAmount > 0)
 		cpsr.carry = (sourceValue >> (shiftAmount - 1) & 1);
 	cpsr.negative = static_cast<int32_t>(result) < 0;
 	cpsr.zero = result == 0;
 }
 
-uint32_t Ror::shift(const uint32_t sourceValue, const uint8_t shiftAmount) {
+uint32_t Ror::shift(const uint32_t sourceValue, const uint32_t shiftAmount) {
 	if (shiftAmount > 32) {
 		Ror::shift(sourceValue, shiftAmount - 32);
 	}
@@ -60,12 +60,12 @@ uint32_t Ror::shift(const uint32_t sourceValue, const uint8_t shiftAmount) {
 }
 
 
-void Rrx::calcConditions(CPSR_t& cpsr, const uint32_t result, const uint32_t sourceValue, const uint8_t shiftAmount)  {
+void Rrx::calcConditions(CPSR_t& cpsr, const uint32_t result, const uint32_t sourceValue, const uint32_t shiftAmount)  {
 	cpsr.carry = sourceValue & 1;
 	cpsr.negative = static_cast<int32_t>(result) < 0;;
 	cpsr.zero = result == 0;
 }
 
-uint32_t Rrx::shift(const uint32_t sourceValue, const uint8_t carry) {
+uint32_t Rrx::shift(const uint32_t sourceValue, const uint32_t carry) {
 	return (carry << 31) | (sourceValue >> 1);
 }
